Automatically Generate XML Files
Implement xml-generator.php that will automatically generate XML files
from the database based on an ID passed in a get query.  This allows
new calls to be added using new MP3s without having to create a new XML
file by hand everytime.  This is the first step to a full web
implementation.